% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readODS.R
\name{get_num_sheet_in_ods}
\alias{getNrOfSheetsInODS}
\alias{get_num_sheet_in_ods}
\title{get the number of sheets in an ods file}
\usage{
get_num_sheet_in_ods(file)

getNrOfSheetsInODS(file)
}
\arguments{
\item{file}{the name of the ods file}
}
\description{
get the number of sheets in an ods file
}
\author{
Chung-hong Chan <chainsawtiney@gmail.com>, Gerrit-Jan Schutten <phonixor@gmail.com>
}
\seealso{
use \code{\link{read_ods}} to read the data
}

