% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ods_sheets.R
\name{get_num_sheets_in_fods}
\alias{get_num_sheets_in_fods}
\title{Get the Number of Sheets in an FODS File}
\usage{
get_num_sheets_in_fods(path, include_external_data = FALSE)
}
\arguments{
\item{path}{path to the fods file}

\item{include_external_data}{A boolean value declaring if sheets holding archived linked data should be included}
}
\value{
Number of sheets
}
\description{
Get the number of sheets in an fods file
}
\examples{
\dontrun{
# Get the number of sheets
get_num_sheets_in_fods("starwars.fods")
}
}
\seealso{
use \code{\link{read_fods}} to read the data
}
\author{
Peter Brohan \href{mailto:peter.brohan+cran@gmail.com}{peter.brohan+cran@gmail.com}, Chung-hong Chan \href{mailto:chainsawtiney@gmail.com}{chainsawtiney@gmail.com}, Gerrit-Jan Schutten \href{mailto:phonixor@gmail.com}{phonixor@gmail.com}
}
