% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readODS.R
\name{get_num_sheets_in_ods}
\alias{get_num_sheets_in_ods}
\alias{getNrOfSheetsInODS}
\title{get the number of sheets in an ods file}
\usage{
get_num_sheets_in_ods(path)

getNrOfSheetsInODS(path)
}
\arguments{
\item{path}{Path to the ods file}
}
\value{
Number of sheets
}
\description{
get the number of sheets in an ods file
}
\seealso{
use \code{\link{read_ods}} to read the data
}
\author{
Chung-hong Chan \href{mailto:chainsawtiney@gmail.com}{chainsawtiney@gmail.com}, Gerrit-Jan Schutten \href{mailto:phonixor@gmail.com}{phonixor@gmail.com}
}
