% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getNrOfSheetsInODS}
\alias{getNrOfSheetsInODS}
\title{getNrOfSheets}
\usage{
getNrOfSheetsInODS(file = NULL)
}
\arguments{
\item{file}{path to the .ods file}
}
\description{
returns the number of sheets in the .ods file
}
\details{
use readODS() to actualy get the sheets
}

