% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lettersToNumber}
\alias{lettersToNumber}
\title{lettersToNumber}
\usage{
lettersToNumber(listOfStrings = NULL)
}
\arguments{
\item{listOfStrings}{the strings you want to convert to numbers}
}
\description{
converts microplate row names and Excel & ODS column names into numbers
}
\details{
supports lists of chars!
ignores case

A=1
Z=26
ZA=27
ZZ=702
AAA=703

lettersToNumber("AAa")
}

