% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ods_sheets.R
\name{list_ods_sheets}
\alias{list_ods_sheets}
\alias{get_num_sheets_in_ods}
\alias{list_fods_sheets}
\alias{get_num_sheets_in_fods}
\title{Get information in an (F)ODS File}
\usage{
list_ods_sheets(path, include_external_data = FALSE)

get_num_sheets_in_ods(path, include_external_data = FALSE)

list_fods_sheets(path, include_external_data = FALSE)

get_num_sheets_in_fods(path, include_external_data = FALSE)
}
\arguments{
\item{path}{Path to the (f)ods file}

\item{include_external_data}{A boolean value to show or hide sheets containing archived linked data (default false)}
}
\value{
For \verb{list_(f)ods_sheets}, a character vector of sheet names; for \verb{get_num_sheets_in_(f)ods}, the total number of sheets
}
\description{
\verb{list_(f)ods_sheets} lists all sheets in an (f)ods file.
}
\examples{
\dontrun{
# Get the list of names of sheets
list_ods_sheets("starwars.ods")
list_fods_sheets("starwars.fods")
# Get the number of sheets
get_num_sheets_in_ods("starwars.ods")
get_num_sheets_in_fods("starwars.fods")
}
}
\seealso{
use \code{\link{read_ods}} to read the data
}
\author{
Peter Brohan \href{mailto:peter.brohan+cran@gmail.com}{peter.brohan+cran@gmail.com}, Chung-hong Chan \href{mailto:chainsawtiney@gmail.com}{chainsawtiney@gmail.com}, Gerrit-Jan Schutten \href{mailto:phonixor@gmail.com}{phonixor@gmail.com}
}
