% Generated by roxygen2 (4.0.1): do not edit by hand
\name{numberToLetters}
\alias{numberToLetters}
\title{numberToLetters}
\usage{
numberToLetters(listOfNumbers = NULL)
}
\arguments{
\item{listOfNumbers}{the numbers you want to convert to chars}
}
\description{
converts numbers to microplate row names and Excel & ODS column names
}
\details{
1=A
26=Z
27=ZA
702=ZZ
703=AAA

supports lists of numbers!

numberToLetters(1:1000)
}
\keyword{internal}

