% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.ods}
\alias{read.ods}
\title{read.ods}
\usage{
read.ods(file = NULL, sheet = NULL, formulaAsFormula = F)
}
\arguments{
\item{file}{filepath to the ods file}

\item{sheet}{select the sheet(s) you want, if left empty it will return all sheets in a list, if only 1 number is given it will return the sheet as a data.frame (not as a list of data.frames)}

\item{formulaAsFormula}{a switch to display formulas as formulas "SUM(A1:A3)" or as the resulting value "3"... or "8"..}
}
\description{
returns a list of data.frames 1 data.frame per sheet
}
\details{
the data.frame contains all strings (not factors)
}

