% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ods.R
\name{readODS_progress}
\alias{readODS_progress}
\title{Determine whether to show progress bar}
\usage{
readODS_progress()
}
\value{
logical indicating whether to show progress
}
\description{
By default, readODS displays a progress bar unless one of the following is TRUE:
\itemize{
\item The progress is explicitly disabled by setting options(readODS.show_progress = FALSE).
\item The code is run in a non-interactive session (interactive() is FALSE).
\item The code is run by knitr / rmarkdown.
}
}
