% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_ods_sheets.R
\name{get_num_sheets_in_ods}
\alias{get_num_sheets_in_ods}
\title{Get the Number of Sheets in an ODS File}
\usage{
get_num_sheets_in_ods(path, include_external_data = FALSE)
}
\arguments{
\item{path}{path to the ods file}

\item{include_external_data}{A boolean value declaring if external data sheets should be counted}
}
\value{
Number of sheets
}
\description{
Get the number of sheets in an ods file
}
\seealso{
use \code{\link{read_ods}} to read the data
}
\author{
Peter Brohan \href{mailto:peter.brohan+cran@gmail.com}{peter.brohan+cran@gmail.com}, Chung-hong Chan \href{mailto:chainsawtiney@gmail.com}{chainsawtiney@gmail.com}, Gerrit-Jan Schutten \href{mailto:phonixor@gmail.com}{phonixor@gmail.com}
}
