% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readODS.R
\name{list_ods_sheets}
\alias{list_ods_sheets}
\alias{ods_sheets}
\title{List all sheets in an ods file.}
\usage{
list_ods_sheets(path)

ods_sheets(path)
}
\arguments{
\item{path}{Path to the ods file}
}
\value{
A character vector of sheet names.
}
\description{
List all sheets in an ods file.
}
