% Generated by roxygen2 (4.0.1): do not edit by hand
\name{odsPreParser}
\alias{odsPreParser}
\title{odsPreParser}
\usage{
odsPreParser(file = NULL)
}
\arguments{
\item{file}{- the xml file to be parsed...}
}
\description{
libre office can do crap like this:
<text:p>></text:p>
which is not valid xml... and the XML package doesn't like that..
OK NM NOW IT DOES! leaving the code here in case of changes...

also to not make the code fulgy as all hell, the content.xml file is first scanned for these XML violations, and then fixed!
<text:p>></text:p> --> <text:p>&gt</text:p>
}
\details{
WTF!!! XML suddenly started being able to handle these files... or maybe just on windows?

also returns an .ods root..
}

