% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readODS.R
\name{read.ods}
\alias{read.ods}
\title{read data from ods files (depreciated)}
\usage{
read.ods(file = NULL, sheet = NULL, formulaAsFormula = F)
}
\arguments{
\item{file}{the name of the ods file which the data are to be read from.}

\item{sheet}{default to NULL with all sheets being read as a list of data.frame. If a number is given, the sheet as a single data.frame is returned.}

\item{formulaAsFormula}{Logical, a switch to display formulas as formulas "SUM(A1:A3)" or as the resulting value "3"... or "8"..}
}
\description{
returns a list of data frames with one data frame per sheet. This is a wrapper to read_ods  for backward compatibile with previous version of readODS. Please use read_ods if possible.
}
\details{
the data.frame contains all strings (not factors)
}
\author{
Chung-hong Chan <chainsawtiney@gmail.com>, Gerrit-Jan Schutten <phonixor@gmail.com>
}

