% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readODS.R
\name{read_ods}
\alias{read_ods}
\title{read data from ods files}
\usage{
read_ods(file = NULL, sheet = 1, header = FALSE,
  formula_as_formula = FALSE, skip = 0)
}
\arguments{
\item{file}{the name of the ods file which the data are to be read from.}

\item{sheet}{numeric, the sheet number to be read from. The default is 1.}

\item{header}{logical,  indicating whether the file contains the names of the variables as its first line.}

\item{formula_as_formula}{logical, a switch to display formulas as formulas "SUM(A1:A3)" or as the resulting value "3"... or "8"..}

\item{skip}{numeric, the number of lines of the data file to skip before beginning to read data.}
}
\value{
a data frame (\code{data.frame}) containing a representation of data in the ods file. All data are read as characters.
}
\description{
Funtion to read a single sheet from a ods file and return a data frame.
}
\author{
Chung-hong Chan <chainsawtiney@gmail.com>
}

