% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readODS-package.R
\docType{package}
\name{readODS-package}
\alias{readODS}
\alias{readODS-package}
\title{readODS: Read and Write ODS Files}
\description{
Read ODS (OpenDocument Spreadsheet) into R as data frame. Also support writing data frame into ODS file.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/readODS}
  \item Report bugs at \url{https://github.com/ropensci/readODS/issues}
}

}
\author{
\strong{Maintainer}: Chung-hong Chan \email{chainsawtiney@gmail.com} (\href{https://orcid.org/0000-0002-6232-7530}{ORCID})

Authors:
\itemize{
  \item Gerrit-Jan Schutten \email{phonixor@gmail.com}
  \item Peter Brohan \email{peter.brohan@gmail.com}
  \item Thomas J. Leeper \email{thosjleeper@gmail.com} (\href{https://orcid.org/0000-0003-4097-6326}{ORCID})
  \item Detlef Steuer \email{steuer@hsu-hh.de} (\href{https://orcid.org/0000-0003-2676-5290}{ORCID})
}

Other contributors:
\itemize{
  \item John Foster \email{john.x.foster@nab.com.au} [contributor]
  \item Sergio Oller [contributor]
  \item Jim Hester \email{jim.hester@rstudio.com} (\href{https://orcid.org/0000-0002-2739-7082}{ORCID}) [contributor]
  \item Stephen Watts [contributor]
  \item Arthur Katossky [contributor]
  \item Stas Malavin [contributor]
  \item Duncan Garmonsway [contributor]
  \item Mehrad Mahmoudian [contributor]
  \item Matt Kerlogue [contributor]
  \item Michal Lauer \email{michal.lauer.25@gmail.com} [contributor]
  \item Till Straube \email{straube@geo.uni-frankfurt.de} [contributor]
}

}
\keyword{internal}
