% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeODS.R
\name{write_ods}
\alias{write_ods}
\title{write data to ods file}
\usage{
write_ods(x, path)
}
\arguments{
\item{x}{a data.frame}

\item{path}{Path to the ods file to write.}
}
\value{
the value of \code{path} invisibly.
}
\description{
Function to write a single data.frame to an ods file.
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
